# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ZOVARK SOC Automation API
  version: 1.2.0
  description: |
    AI-powered SOC automation platform. Manages security investigations,
    task orchestration, SIEM integration, shadow mode automation, kill switch
    controls, token quota management, MCP approval gates, analytics,
    and multi-tenant operations.

    ## Authentication
    - **Access tokens** expire in 15 minutes and are returned in the JSON response body.
    - **Refresh tokens** are stored in httpOnly cookies (7-day lifetime) and are NOT sent as Bearer tokens.
      Use `POST /api/v1/auth/refresh` to obtain a new access token.
    - **API keys** (zovark_ prefix) can be used for M2M authentication via X-API-Key header.
  contact:
    name: ZOVARK Platform
  license:
    name: Proprietary
servers:
  - url: http://localhost:8090
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /api/v1/auth/login, /api/v1/auth/callback,
        or /api/v1/auth/refresh. Access tokens expire in 15 minutes. Refresh tokens
        are delivered via httpOnly cookies and cannot be sent as Bearer tokens — use
        the /api/v1/auth/refresh endpoint to rotate access tokens silently.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for M2M authentication (zovark_ prefix)
...
paths:
...
  /api/v1/mcp/approvals/request:
    post:
      operationId: requestMCPApproval
      summary: Request MCP workflow approval
      tags: [MCP Approvals]
      description: |
        Submits a new approval request for an MCP workflow action. Returns
        an approval token that can be polled via the check endpoint. Any
        authenticated user may create a request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, context]
              properties:
                action:
                  type: string
                  description: The action requiring approval
                context:
                  type: object
                  description: Contextual data for the approver
                ttl_seconds:
                  type: integer
                  description: Time-to-live before auto-expiry
                  default: 3600
      responses:
        '201':
          description: Approval request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  approval_id:
                    type: string
                  token:
                    type: string
                  status:
                    type: string
                    example: pending
                  expires_at:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
...
  /api/v1/mcp/approvals/pending:
    get:
      operationId: listMCPApprovals
      summary: List pending MCP approvals (admin)
      tags: [MCP Approvals]
      description: |
        Returns all pending MCP approval requests for the tenant. Admin only.
      responses:
        '200':
          description: Pending MCP approvals
          content:
            application/json:
              schema:
                type: object
                properties:
                  approvals:
                    type: array
                    items:
                      type: object
                      properties:
                        approval_id:
                          type: string
                        token:
                          type: string
                        action:
                          type: string
                        context:
                          type: object
                        status:
                          type: string
                        requested_by:
                          type: string
                        requested_at:
                          type: string
                          format: date-time
                        expires_at:
                          type: string
                          format: date-time
                  count:
                    type: integer
        '401':
          description: Not authenticated
        '403':
          description: Admin role required
...